requirements.txt
aiogram==3.0.0b7
asyncpg==0.27.0
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
secure-smtplib
python-dotenv==1.0.0
openai==0.27.0

